refactor: modify folder structure
