+++ src/lib/index.ts
@@ -1,0 +1,1 @@
+export { default as useJuicyModal } from './hooks/useModal'
